core(devtoolsLog): clarify messagelog separation
