+++ tsconfig.json
@@ -48,11 +48,11 @@
     "lighthouse-core/test/config/config-test.js",
     "lighthouse-core/test/config/default-config-test.js",
     "lighthouse-core/test/create-test-trace.js",
-    "lighthouse-core/test/gather/devtools-log-test.js",
     "lighthouse-core/test/gather/driver/wait-for-condition-test.js",
     "lighthouse-core/test/gather/gatherers/accessibility-test.js",
     "lighthouse-core/test/gather/gatherers/cache-contents-test.js",
     "lighthouse-core/test/gather/gatherers/console-messages-test.js",
+    "lighthouse-core/test/gather/gatherers/devtools-log-test.js",
     "lighthouse-core/test/gather/gatherers/dobetterweb/optimized-images-test.js",
     "lighthouse-core/test/gather/gatherers/dobetterweb/password-inputs-with-prevented-paste-test.js",
     "lighthouse-core/test/gather/gatherers/dobetterweb/response-compression-test.js",
